修改P2P聊天消息存储到字典里

+++ IMiPhone/Persistence/Database/DBChatMessage.h
@@ -16,7 +16,7 @@
 @property (nonatomic) NSInteger targetId;
 @property (nonatomic) NSInteger msgType;
 @property (nonatomic, retain) NSString *content;
-@property (nonatomic) NSInteger mid;
+@property (nonatomic) long mid;
 @property (nonatomic) NSInteger senderUid;
 @property (nonatomic, retain) NSString *sendTime;
